HHH-10707 Proposed minimalist fix.
Converts Hibernate to cache only regions not access strategies.
All regions are cached in a single map and Hibernate throws when
asked to share a region between either incompatible region types
or 'incompatible' CacheDataDescriptions.  Beyond this it is left
up to the cache implementation to complain if it is asked to
access a region in a combination of ways that it doesn't like.

One interesting observation here is that this breaks a bunch of
the Infinispan tests.  It seems that although Infinispan doesn't
tolerate calling buildAccessStrategy(...) multiples times, it's
tests do precisely that.  It appears they only pass in master due
to the access strategy caching being performed by Hibernate.  I
believe this also means the current Infinispan tests are not
currently providing coverage for all the access strategies.